+++ internal/dataplane/translator/translate_cacerts.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/kong/go-kong/kong"
+	"github.com/samber/lo"
 	corev1 "k8s.io/api/core/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -20,26 +21,56 @@ import (
 // getCACerts translates CA certificates Secrets to kong.CACertificates. It ensures every certificate's structure and
 // validity. It skips Secrets that do not contain a valid certificate and reports translation failures for them.
 func (t *Translator) getCACerts() []kong.CACertificate {
-	caCertSecrets, err := t.storer.ListCACerts()
+	caCertSecrets, caCertConfigMaps, err := t.storer.ListCACerts()
 	if err != nil {
 		t.logger.Error(err, "Failed to list CA certs")
 		return nil
 	}
 
-	caCerts := make([]kong.CACertificate, 0, len(caCertSecrets))
-	for _, certSecret := range caCertSecrets {
-		idBytes, ok := certSecret.Data["id"]
+	// intermediateT stores the object and its data for easier handling.
+	// intermediateT.data can come either from secret.Data (map[string]byte) or configmap.Data (map[string]string).
+	type intermediateT struct {
+		obj  client.Object
+		data map[string]string
+	}
+
+	caCertsData := lo.Map(caCertSecrets, func(secret *corev1.Secret, _ int) intermediateT {
+		return intermediateT{
+			obj: secret.DeepCopy(),
+			data: lo.MapValues(secret.Data, func(v []byte, _ string) string {
+				return string(v)
+			}),
+		}
+	})
+	caCertsData = append(caCertsData, lo.Map(caCertConfigMaps, func(configmap *corev1.ConfigMap, _ int) intermediateT {
+		return intermediateT{obj: configmap.DeepCopy(), data: configmap.Data}
+	})...)
+
+	caCerts := make([]kong.CACertificate, 0, len(caCertSecrets)+len(caCertConfigMaps))
+	for _, caCertData := range caCertsData {
+		secretID, ok := caCertData.data["id"]
 		if !ok {
-			t.registerTranslationFailure("invalid CA certificate: missing 'id' field in data", certSecret)
+			t.registerTranslationFailure("invalid secret CA certificate: missing 'id' field in data", caCertData.obj)
 			continue
 		}
-		secretID := string(idBytes)
 
-		caCert, err := toKongCACertificate(certSecret, secretID)
+		// Allow the certificate key to be named either "cert" or "ca.crt".
+		caCertStr, certExists := caCertData.data["cert"]
+		if !certExists {
+			caCertStr, certExists = caCertData.data["ca.crt"]
+			if !certExists {
+				relatedObjects := getPluginsAssociatedWithCACertSecret(secretID, t.storer)
+				relatedObjects = append(relatedObjects, caCertData.obj)
+				t.registerTranslationFailure(fmt.Sprintf(`invalid secret CA certificate %s/%s, neither "cert" nor "ca.crt" key exist`, caCertData.obj.GetNamespace(), caCertData.obj.GetName()), relatedObjects...)
+				continue
+			}
+		}
+
+		caCert, err := toKongCACertificate([]byte(caCertStr), caCertData.obj, secretID)
 		if err != nil {
 			relatedObjects := getPluginsAssociatedWithCACertSecret(secretID, t.storer)
-			relatedObjects = append(relatedObjects, certSecret.DeepCopy())
-			t.registerTranslationFailure(fmt.Sprintf("invalid CA certificate: %s", err), relatedObjects...)
+			relatedObjects = append(relatedObjects, caCertData.obj)
+			t.registerTranslationFailure(fmt.Sprintf("invalid secret CA certificate: %s", err), relatedObjects...)
 			continue
 		}
 
@@ -49,12 +80,8 @@ func (t *Translator) getCACerts() []kong.CACertificate {
 	return caCerts
 }
 
-func toKongCACertificate(certSecret *corev1.Secret, secretID string) (kong.CACertificate, error) {
-	caCertbytes, certExists := certSecret.Data["cert"]
-	if !certExists {
-		return kong.CACertificate{}, errors.New("missing 'cert' field in data")
-	}
-	pemBlock, _ := pem.Decode(caCertbytes)
+func toKongCACertificate(caCertBytes []byte, object client.Object, secretID string) (kong.CACertificate, error) {
+	pemBlock, _ := pem.Decode(caCertBytes)
 	if pemBlock == nil {
 		return kong.CACertificate{}, errors.New("invalid PEM block")
 	}
@@ -71,8 +98,8 @@ func toKongCACertificate(certSecret *corev1.Secret, secretID string) (kong.CACer
 
 	return kong.CACertificate{
 		ID:   kong.String(secretID),
-		Cert: kong.String(string(caCertbytes)),
-		Tags: util.GenerateTagsForObject(certSecret),
+		Cert: kong.String(string(caCertBytes)),
+		Tags: util.GenerateTagsForObject(object),
 	}, nil
 }